mat-dialog and snackBar added

+++ src/app/shared/components/studentlist/studentlist.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { StudentService } from '../../services/student.service';
 import { Istudent } from '../../models/student';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { GetConfirmComponent } from '../get-confirm/get-confirm.component';
+import { SnackbarService } from '../../services/snackbar.service';
 
 @Component({
   selector: 'app-studentlist',
@@ -9,7 +12,11 @@ import { Istudent } from '../../models/student';
 })
 export class StudentlistComponent implements OnInit {
   studentData!: Array<Istudent>;
-  constructor(private _studentService: StudentService) {}
+  constructor(
+    private _studentService: StudentService,
+    private _matDialog: MatDialog,
+    private _snackBar : SnackbarService
+  ) {}
 
   ngOnInit(): void {
     this.studentData = this._studentService.fetchAllStudent();
@@ -19,6 +26,18 @@ export class StudentlistComponent implements OnInit {
   }
 
   onRemove(student: Istudent) {
-    this._studentService.removeStudent(student);
+    let matDialogConfig = new MatDialogConfig();
+    matDialogConfig.width = '600px';
+    matDialogConfig.disableClose = true;
+    let matDialogRef = this._matDialog.open(
+      GetConfirmComponent,
+      matDialogConfig
+    );
+    matDialogRef.afterClosed().subscribe((res) => {
+      if (res === true) {
+        this._studentService.removeStudent(student);
+        this._snackBar.openSnackBar(`${student.fullName} removed successfully`)
+      }
+    });
   }
 }
